feat(project) generic - inclusão do objeto console

+++ src/main/java/com/projeto/universojogos/adapter/outbound/ConsoleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConsoleRepository extends JpaRepository<ConsoleEntity, Integer> {
+public interface ConsoleRepository extends JpaRepository<ConsoleEntity, Long> {
 }
